Update Main.java
Data type changed

+++ src/Main.java
@@ -4,10 +4,10 @@ public class Main {
     public static Scanner in = new Scanner(System.in);
     public static PrintStream out = System.out;
     public static void main(String[] args) {
-        double a = in.nextDouble();
-        double b = in.nextDouble();
-        double c = in.nextDouble();
-        double d = in.nextDouble();
+        int a = in.nextInt();
+        int b = in.nextInt();
+        int c = in.nextInt();
+        int d = in.nextInt();
         int k = 4;
         if (a == b || a == c || a == d) k--;
         if (b == c || b == d) k--;
